refactor: #115 Admin authority check for categoryThird Delete
카테고리(소) 삭제 시 권한 확인하는 기능을 추가하였습니다.

+++ src/main/java/com/akatsuki/pioms/categoryThird/service/CategoryThirdServiceImpl.java
@@ -113,30 +113,30 @@ public class CategoryThirdServiceImpl implements CategoryThirdService{
 
     @Override
     @Transactional
-    public String deleteCategoryThird(int categoryThirdCode/*, int requesterAdminCode*/) {
-//        Optional<Admin> requestorAdmin = adminRepository.findById(requesterAdminCode);
-//        if (requestorAdmin.isEmpty() || requestorAdmin.get().getAdminCode() != 1) {
-//            return ResponseEntity.status(403).body("카테고리 삭제는 루트 관리자만 가능합니다.");
-//        }
+    public ResponseEntity<String> deleteCategoryThird(int categoryThirdCode, int requesterAdminCode) {
+        Optional<Admin> requestorAdmin = adminRepository.findById(requesterAdminCode);
+        if (requestorAdmin.isEmpty() || requestorAdmin.get().getAdminCode() != 1) {
+            return ResponseEntity.status(403).body("카테고리 삭제는 루트 관리자만 가능합니다.");
+        }
 
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
         String formattedDateTime = LocalDateTime.now().format(formatter);
 
         CategoryThird categoryThird = categoryThirdRepository.findByCategoryThirdCode(categoryThirdCode);
         if (categoryThird == null) {
-            return categoryThirdCode + "번 카테고리(소) 카테고리가 없습니다!";
+            return ResponseEntity.badRequest().body(categoryThirdCode + "번 카테고리(소) 카테고리가 없습니다!");
         }
 
         // Check if there are any products associated with this categoryThirdCode
         List<Product> products = productRepository.findByCategoryThird_CategoryThirdCode(categoryThirdCode);
         if (!products.isEmpty()) {
-            return "상품이 존재하는 해당 " + categoryThirdCode + "번 카테고리(소) 카테고리는 삭제할 수 없습니다!";
+            return ResponseEntity.badRequest().body("상품이 존재하는 해당 " + categoryThirdCode + "번 카테고리(소) 카테고리는 삭제할 수 없습니다!");
         }
 
         categoryThird.setCategoryThirdDeleteDate(formattedDateTime);
         categoryThirdRepository.delete(categoryThird);
         logService.saveLog("root", LogStatus.삭제,categoryThird.getCategoryThirdName(),"CategoryThird");
-        return categoryThirdCode + "번의 해당 카테고리(소) 카테고리가 성공적으로 삭제되었습니다!";
+        return ResponseEntity.badRequest().body(categoryThirdCode + "번의 해당 카테고리(소) 카테고리가 성공적으로 삭제되었습니다!");
     }
 
 //    @Override
